Increase `Duration` precision to 64-bits

+++ NAS2D/Timer.h
@@ -32,15 +32,15 @@ namespace NAS2D
 	 *
 	 * A static method is provided for the raw time in ticks.
 	 *
-	 * With the current implementation, raw ticks are since app startup, and wrap back to 0 after about 49 days.
+	 * Raw ticks are since app startup.
 	 */
 	class Timer
 	{
 	public:
-		static uint32_t tick();
+		static uint64_t tick();
 
 		Timer();
-		Timer(uint32_t startTick);
+		Timer(uint64_t startTick);
 
 		Timer(const Timer&) = default;
 		Timer& operator=(const Timer&) = default;
@@ -51,7 +51,7 @@ namespace NAS2D
 		void reset();
 
 	private:
-		uint32_t mStartTick;
+		uint64_t mStartTick;
 	};
 
 } // namespace
